Запрос на списание баллов лояльности

+++ internal/model/accrual_info.go
@@ -5,6 +5,6 @@ import "time"
 type AccrualInfo struct {
 	OrderNumber string    `json:"order"`
 	Status      string    `json:"status"`
-	Sum         int       `json:"accrual,omitempty"`
+	Sum         float64   `json:"accrual,omitempty"`
 	CreatedAt   time.Time `json:"uploaded_at,omitempty"`
 }
